remove dead code add as module in zed
